+++ cypress/e2e/plugins/2-prototype-kit-plugin-tests/remove-govuk-frontend.cypress.js
@@ -1,11 +1,20 @@
 const { managePluginsPagePath, performPluginAction } = require('../plugin-utils')
 const { uninstallPlugin, installPlugin } = require('../../utils')
+const path = require('path')
 
 const plugin = 'govuk-frontend'
 const pluginName = 'GOV.UK Frontend'
 const dependentPlugin = '@govuk-prototype-kit/common-templates'
+const appConfigPath = path.join('app', 'config.json')
 
 describe('Manage prototype pages without govuk-frontend', () => {
+  before(() => {
+    cy.task('copyFromStarterFiles', { filename: appConfigPath })
+    cy.task('addToConfigJson', { allowGovukFrontendUninstall: true })
+  })
+  after(() => {
+    cy.task('copyFromStarterFiles', { filename: appConfigPath })
+  })
   it('Uninstall govuk-frontend', () => {
     uninstallPlugin(dependentPlugin)