+++ CHANGELOG.md
@@ -15,8 +15,10 @@
 - [#2150: Prevent CSURF deprecated warning](https://github.com/alphagov/govuk-prototype-kit/pull/2150)
 
 - [#2130: Create our own file store functionality in place of the session-file-store package](https://github.com/alphagov/govuk-prototype-kit/pull/2130)
-- 
+
 - [#2173: Autostore should not run on management pages](https://github.com/alphagov/govuk-prototype-kit/pull/2173)
+
+- [#2172: Hiding govuk-frontend uninstall until we deal with dependent plugins](https://github.com/alphagov/govuk-prototype-kit/pull/2172)
 
 ## 13.6.2